+++ sb3_contrib/cem/policies.py
@@ -1,0 +1,4 @@
+from sb3_contrib.common.policies import ESLinearPolicy, ESPolicy
+
+MlpPolicy = ESPolicy
+LinearPolicy = ESLinearPolicy